// src/constants/voice-state-types.ts
export enum VoiceStateEventType {
    JOIN = "JOIN",
    LEAVE = "LEAVE",
    MOVE = "MOVE",
    STREAM_START = "STREAM_START",
    STREAM_STOP = "STREAM_STOP",
    SELF_MUTE = "SELF_MUTE",
    SELF_UNMUTE = "SELF_UNMUTE",
    SELF_DEAF = "SELF_DEAF",
    SELF_UNDEAF = "SELF_UNDEAF",
    SERVER_MUTE = "SERVER_MUTE",
    SERVER_UNMUTE = "SERVER_UNMUTE",
    SERVER_DEAF = "SERVER_DEAF",
    SERVER_UNDEAF = "SERVER_UNDEAF",
    VIDEO_START = "VIDEO_START",
    VIDEO_STOP = "VIDEO_STOP",
    SUPPRESS = "SUPPRESS",
    UNSUPPRESS = "UNSUPPRESS"
}